Added memory specification
This commit adds the --mem option to allow simple memory requests per job, and
a --memvars option to specify which variables are set with that value and
passed on during job submission. For example,

    qbatch --mem 1G --memvars h_vmem,vf commands.txt

For SGE, this would pass an -l argument to qsub like so:

    -l h_vmem=1G,vf=1G

Environment variables are also exposed to set default values for both of these
options.

This fixes issue #5