Checks presence of queue binary before submitting jobs
This change fixes #177 where running --help with the QBATCH_SYSTEM
environment variable set, but no queuing system binary in the path, will
cause qbatch to exit with an error message but not display help.

A test case is provided here that does not quite follow unit testing
conventions in that it does teardown within the function via
try..finally. I opted for this because it seemed less unwieldy compared
to importing the nose module and attaching a separate teardown function.